requirements.txt
arrow==1.3.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
binaryornot==0.4.4
black==24.2.0
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
channels==4.0.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
cookiecutter==2.6.0
cryptography==42.0.5
daphne==4.1.0
distlib==0.3.8
Django==4.2.11
django-browser-reload==1.12.1
django-tailwind==3.6.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
filelock==3.13.1
hyperlink==21.0.0
identify==2.5.35
idna==3.6
incremental==22.10.0
isort==5.13.2
Jinja2==3.1.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
pre-commit==3.6.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.17.2
PyJWT==1.7.1
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
redis==5.0.3
requests==2.31.0
rich==13.7.1
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
tomli==2.0.1
Twisted==24.3.0
txaio==23.1.1
types-python-dateutil==2.8.19.20240311
typing_extensions==4.10.0
urllib3==2.2.1
virtualenv==20.25.1
zope.interface==6.2
